validation function
I have started work on validation. It may be working, but it needs
some improvements.

honestly, I am not so about this project right now. I have decided that
this won't be working perfect, so I just want to finish it and don't
touch it in the near future. it is not so fun working on something that
I won't probably use ever